Разбил один контроллер на несколько составляющих. Немного изменил формы.
