+++ admin-ui/src/main/java/ru/geekbrains/service/ProductServiceImpl.java
@@ -63,7 +63,7 @@ public class ProductServiceImpl implements ProductService, Serializable {
         product.setCategory(productRepr.getCategory());
         product.setBrand(productRepr.getBrand());
         product.setPrice(productRepr.getPrice());
-        if (productRepr.getNewPictures() != null) {
+        if (!productRepr.getNewPictures()[0].isEmpty()) {
             for (MultipartFile newPicture : productRepr.getNewPictures()) {
                 logger.info("Product {} file {} size {} contentType {}", productRepr.getId(),
                         newPicture.getOriginalFilename(), newPicture.getSize(), newPicture.getContentType());
