+++ admin-ui/src/main/java/ru/geekbrains/service/UserService.java
@@ -1,0 +1,10 @@
+package ru.geekbrains.service;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.domain.Specification;
+import ru.geekbrains.entity.User;
+import ru.geekbrains.repr.UserRepr;
+
+public interface UserService {
+}
