+++ admin-ui/src/main/resources/templates/editbrand.html
@@ -23,7 +23,7 @@
                     </table>
                 </div>
                 <div class="col-12">
-                    <form th:action="@{/updatebrand}" method="post" th:object="${brand}">
+                    <form th:action="@{/brands/update}" method="post" th:object="${brand}">
                         <input type="hidden" th:field="*{id}">
                         <div class="form-group">
                             <input type="text" class="form-control" th:field="*{name}" placeholder="Enter new name">
