+++ admin-ui/src/main/resources/templates/editcategory.html
@@ -23,7 +23,7 @@
                     </table>
                 </div>
                 <div class="col-12">
-                    <form th:action="@{/updatecategory}" method="post" th:object="${category}">
+                    <form th:action="@{/categories/update}" method="post" th:object="${category}">
                         <input type="hidden" th:field="*{id}">
                         <div class="form-group">
                             <input type="text" class="form-control" th:field="*{name}" placeholder="Enter new name">
