+++ admin-ui/src/main/resources/templates/editproduct.html
@@ -37,7 +37,7 @@
                     </div>
                 </div>
                 <div class="col-12">
-                    <form th:action="@{/updateproduct}" method="post" th:object="${product}" enctype="multipart/form-data">
+                    <form th:action="@{/products/update}" method="post" th:object="${product}" enctype="multipart/form-data">
                         <input type="hidden" th:field="*{id}">
                         <div class="form-group">
                             <input type="text" class="form-control" th:field="*{name}" placeholder="Enter new name">
