+++ admin-ui/src/main/resources/templates/fragments/sidebar.html
@@ -9,8 +9,8 @@
                 <i class="fa fa-fw fa-link"></i> Users
             </a>
             <ul id="sm_expand_1" class="list-unstyled collapse">
-                <li><a th:href="@{/users}">All users</a></li>
-                <li><a th:href="@{/newuser}">Create user</a></li>
+                <li><a th:href="@{/users/all}">All users</a></li>
+                <li><a th:href="@{/users/new}">Create user</a></li>
             </ul>
         </li>
         <li>
@@ -18,8 +18,8 @@
                 <i class="fa fa-fw fa-link"></i> Roles
             </a>
             <ul id="sm_expand_2" class="list-unstyled collapse">
-                <li><a th:href="@{/roles}">All roles</a></li>
-                <li><a th:href="@{/newrole}">Create role</a></li>
+                <li><a th:href="@{/roles/all}">All roles</a></li>
+                <li><a th:href="@{/roles/new}">Create role</a></li>
             </ul>
         </li>
         <li>
@@ -27,8 +27,8 @@
                 <i class="fa fa-fw fa-link"></i> Brands
             </a>
             <ul id="sm_expand_3" class="list-unstyled collapse">
-                <li><a th:href="@{/brands}">All brands</a></li>
-                <li><a th:href="@{/newbrand}">Create brand</a></li>
+                <li><a th:href="@{/brands/all}">All brands</a></li>
+                <li><a th:href="@{/brands/new}">Create brand</a></li>
             </ul>
         </li>
         <li>
@@ -36,8 +36,8 @@
                 <i class="fa fa-fw fa-link"></i> Categories
             </a>
             <ul id="sm_expand_4" class="list-unstyled collapse">
-                <li><a th:href="@{/categories}">All categories</a></li>
-                <li><a th:href="@{/newcategory}">Create category</a></li>
+                <li><a th:href="@{/categories/all}">All categories</a></li>
+                <li><a th:href="@{/categories/new}">Create category</a></li>
             </ul>
         </li>
         <li>
@@ -45,8 +45,8 @@
                 <i class="fa fa-fw fa-link"></i> Products
             </a>
             <ul id="sm_expand_5" class="list-unstyled collapse">
-                <li><a th:href="@{/products}">All products</a></li>
-                <li><a th:href="@{/newproduct}">Create product</a></li>
+                <li><a th:href="@{/products/all}">All products</a></li>
+                <li><a th:href="@{/products/new}">Create product</a></li>
             </ul>
         </li>
     </ul>
