+++ admin-ui/src/main/resources/templates/products.html
@@ -34,8 +34,8 @@
                             <td th:text="${product.brand}">Brand</td>
                             <td th:text="${product.category}">Category</td>
                             <td>
-                                <a class="btn btn-success" th:href="@{editproduct(id=${product.id})}"><i class="fas fa-edit"></i> </a>
-                                <a class="btn btn-danger" th:href="@{deleteproduct(id=${product.id})}"><i class="far fa-trash-alt"></i> </a>
+                                <a class="btn btn-success" th:href="@{/products/edit(id=${product.id})}"><i class="fas fa-edit"></i> </a>
+                                <a class="btn btn-danger" th:href="@{/products/delete(id=${product.id})}"><i class="far fa-trash-alt"></i> </a>
                             </td>
                         </tr>
                         </tbody>
@@ -48,7 +48,7 @@
                                     th:classappend="${num == products.number + 1} ? 'active'">
                                     <a class="page-link"
                                        th:text="${num}"
-                                       th:href="@{products(page=${num}, size=${products.size}, name=${param.name})}">1</a>
+                                       th:href="@{/products/all(page=${num}, size=${products.size}, name=${param.name})}">1</a>
                                 </li>
                             </ul>
                         </nav>
